refactor scheduler and recursion tip
